+++ common/src/main/java/traben/solid_mobs/config/SolidMobsCommands.java
@@ -1,24 +1,26 @@
 package traben.solid_mobs.config;
 
+import com.mojang.brigadier.Command;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.BoolArgumentType;
 import com.mojang.brigadier.arguments.FloatArgumentType;
 import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandExceptionType;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import net.minecraft.command.CommandRegistryAccess;
 import net.minecraft.entity.EntityType;
+import net.minecraft.registry.Registries;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
+import org.jetbrains.annotations.NotNull;
 import traben.solid_mobs.SolidMobsCrossPlatformHelper;
 import traben.solid_mobs.SolidMobsMain;
 
-import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.function.Consumer;
 
@@ -33,57 +35,45 @@ public class SolidMobsCommands {
         return new CommandSyntaxException(exception, () -> str);
     }
 
-    public static int blacklistAdd(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
+    public static int listAdd(CommandContext<ServerCommandSource> context, Set<String> list, String name) throws CommandSyntaxException {
         try {
             String add = context.getArgument("text", String.class);
-            Set<String> list = new HashSet<>(List.of(SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist));
             list.add(add);
-            SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist = list.toArray(new String[0]);
             SolidMobsMain.saveConfig();
             SolidMobsMain.resetExemptions();
             context.getSource().getServer().getPlayerManager().getPlayerList().forEach(SolidMobsCrossPlatformHelper::sendConfigToClient);
-            sendCommandFeedback(context, (" [" + add + "] has been added to the collision blacklist."));
+            sendCommandFeedback(context, (" [" + add + "] has been added to the " + name));
             return 1;
         } catch (Exception e) {
             throw commandException(e.getMessage());
         }
     }
 
-    public static int blacklistRemove(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
+    public static int listRemove(CommandContext<ServerCommandSource> context, Set<String> list, String name) throws CommandSyntaxException {
         try {
             String remove = context.getArgument("text", String.class);
-            Set<String> list = new HashSet<>(List.of(SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist));
             list.remove(remove);
-            SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist = list.toArray(new String[0]);
             SolidMobsMain.saveConfig();
             SolidMobsMain.resetExemptions();
             context.getSource().getServer().getPlayerManager().getPlayerList().forEach(SolidMobsCrossPlatformHelper::sendConfigToClient);
-            sendCommandFeedback(context, (" [" + remove + "] has been removed from the collision blacklist."));
+            sendCommandFeedback(context, (" [" + remove + "] has been removed from the " + name));
             return 1;
         } catch (Exception e) {
             throw commandException(e.getMessage());
         }
     }
 
-    public static int blacklistClear(CommandContext<ServerCommandSource> context) {
-        String[] held = SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist;
-        SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist = new String[0];
+    public static int listClear(CommandContext<ServerCommandSource> context, Set<String> list, String name) {
+        Object held = new HashSet<>(list);
+        list.clear();
         SolidMobsMain.saveConfig();
         SolidMobsMain.resetExemptions();
         context.getSource().getServer().getPlayerManager().getPlayerList().forEach(SolidMobsCrossPlatformHelper::sendConfigToClient);
-        sendCommandFeedback(context, ("The collision blacklist has been cleared, it was: " + Arrays.toString(held)));
+        sendCommandFeedback(context, ("The "+name+" has been cleared, it was: " + held));
         return 1;
     }
 
-    public static int blacklistList(CommandContext<ServerCommandSource> context) {
-        sendCommandFeedback(context, (
-                "The HARDCODED collision blacklist (set by the settings or the dev): "
-                        + SolidMobsMain.EXEMPT_ENTITIES.toString()
-                        + "\n§7_____________________________§r\nThe Custom collision blacklist (set by you): "
-                        + Arrays.toString(SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist)
-        ));
-        return 1;
-    }
+
 
 
     public static int recentCollisionsQueryAll(CommandContext<ServerCommandSource> context) {
@@ -195,11 +185,11 @@ public class SolidMobsCommands {
     private static int help(CommandContext<ServerCommandSource> context) {
         sendCommandFeedback(context, ("""
                 §c§lSolids Mobs, command help:§r
-                           
+                
                 §a§o~listRecentCollisions§r - Displays a list of recent entity collisions the mod has been involved with, and their outcomes.
-                           
+                
                 §a§o~entityBlacklist§r  - Allows modification of the entity blacklist (entities that will not be modified by the mod). Entity names for these commands can be found using "getRecentCollisions".
-                           
+                
                 §a§o~settings <setting>§r  - Allows modification of all server side settings. Must be a server Operator or be in Single-Player. Each setting will explain what is does if you run it without setting a value."""));
         return 1;
     }
@@ -219,33 +209,32 @@ public class SolidMobsCommands {
                         .then(CommandManager.literal("resultCollision").executes(SolidMobsCommands::recentCollisionsQueryYes))
                         .then(CommandManager.literal("resultNoCollision").executes(SolidMobsCommands::recentCollisionsQueryNo))
                 )
-                .then(CommandManager.literal("entityBlacklist")
-                        .executes(SolidMobsCommands::blacklistList)
-                        .then(CommandManager.literal("list").executes(SolidMobsCommands::blacklistList))
-                        .then(CommandManager.literal("add").then(CommandManager.argument("text", StringArgumentType.string()).executes(SolidMobsCommands::blacklistAdd)))
-                        .then(CommandManager.literal("remove").then(CommandManager.argument("text", StringArgumentType.string()).executes(SolidMobsCommands::blacklistRemove)))
-                        .then(CommandManager.literal("clear").executes(SolidMobsCommands::blacklistClear))
-                )
-
+                .then(entityListCommand("entityBlacklist", SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist, blacklistList()))
+                .then(entityListCommand("entityBounceList", SolidMobsMain.solidMobsConfigData.entityBounceList, bounceListList()))
                 .then(CommandManager.literal("settings")
                         .executes(SolidMobsCommands::printSettings)
-                        .then(CommandManager.literal("slimesAreBouncy").executes((context ->
-                                        returnMessageValue(context, SolidMobsMain.solidMobsConfigData.bouncySlimes,
-                                                "Sets whether or not Slimes & Magma cubes cause you to bounce like the blocks.")))
-                                .then(CommandManager.argument("set", BoolArgumentType.bool())
-                                        .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.bouncySlimes = val))))
-                        )
                         .then(CommandManager.literal("platformMode").executes((context ->
                                         returnMessageValue(context, SolidMobsMain.solidMobsConfigData.platformMode,
                                                 "If 'true' mobs will only be solid when walking on top of them.\n You can also press crouch to fall through them like platforms.")))
                                 .then(CommandManager.argument("set", BoolArgumentType.bool())
                                         .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.platformMode = val))))
+                        ).then(CommandManager.literal("playerOnlyMode").executes((context ->
+                                        returnMessageValue(context, SolidMobsMain.solidMobsConfigData.playerOnlyMode,
+                                                "If 'true' mobs will only be solid when colliding with a player not other mobs.")))
+                                .then(CommandManager.argument("set", BoolArgumentType.bool())
+                                        .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.playerOnlyMode = val))))
                         )
                         .then(CommandManager.literal("canShoveMobs").executes((context ->
                                         returnMessageValue(context, SolidMobsMain.solidMobsConfigData.allowShovingMobs,
                                                 "Sets whether or not you can shove mobs, mobs can be shoved by crouching and right clicking them.\nThis allows you to get un-stuck when crammed in by a mob.\nBeware neutral mobs may get mad after being shoved.\nThe cooldown between shoving attempts can be set with the 'shovingMobsCoolDownTicks' setting")))
                                 .then(CommandManager.argument("set", BoolArgumentType.bool())
                                         .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.allowShovingMobs = val))))
+                        )
+                        .then(CommandManager.literal("ignoreCollisionsWhenCrouching").executes((context ->
+                                        returnMessageValue(context, SolidMobsMain.solidMobsConfigData.ignoreCollisionsWhenCrouching,
+                                                "Allows you to ignore collisions when crouching. Reverts to vanilla behaviour.")))
+                                .then(CommandManager.argument("set", BoolArgumentType.bool())
+                                        .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.ignoreCollisionsWhenCrouching = val))))
                         )
                         .then(CommandManager.literal("canItemsCollide").executes((context ->
                                         returnMessageValue(context, SolidMobsMain.solidMobsConfigData.allowItemCollisions,
@@ -270,6 +259,12 @@ public class SolidMobsCommands {
                                                 "Sets whether or not Villagers can collide.")))
                                 .then(CommandManager.argument("set", BoolArgumentType.bool())
                                         .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.allowVillagerCollisions = val))))
+                        )
+                        .then(CommandManager.literal("canBabiesCollide").executes((context ->
+                                        returnMessageValue(context, SolidMobsMain.solidMobsConfigData.allowBabyCollisions,
+                                                "Sets whether or not Baby mobs can collide.")))
+                                .then(CommandManager.argument("set", BoolArgumentType.bool())
+                                        .executes((context -> runBooleanConfigCommand(context, (val) -> SolidMobsMain.solidMobsConfigData.allowBabyCollisions = val))))
                         )
                         .then(CommandManager.literal("canInvisibleMobsCollide").executes((context ->
                                         returnMessageValue(context, SolidMobsMain.solidMobsConfigData.allowInvisibleCollisions,
@@ -305,5 +300,51 @@ public class SolidMobsCommands {
                         .then(CommandManager.literal("listAllSettings").executes(SolidMobsCommands::printSettings))
                 )
         );
+    }
+
+    private static LiteralArgumentBuilder<ServerCommandSource> entityListCommand(String name, Set<String> list, Command<ServerCommandSource> listPrinter) {
+        return CommandManager.literal(name)
+                .executes(listPrinter)
+                .then(CommandManager.literal("list")
+                        .executes(listPrinter))
+                .then(CommandManager.literal("add")
+                        .then(CommandManager.argument("text", StringArgumentType.string())
+                                .executes(context -> listAdd(context, list, name))))
+                .then(CommandManager.literal("remove")
+                        .then(CommandManager.argument("text", StringArgumentType.string())
+                                .executes(c-> listRemove(c, list, name))))
+                .then(CommandManager.literal("clear")
+                        .executes(context -> listClear(context, list, name)))
+                .then(CommandManager.literal("add_all_entities")
+                        .executes(c -> addAllToList(c, list, name)));
+    }
+
+    private static int addAllToList(CommandContext<ServerCommandSource> context, Set<String> list, String name) {
+        int size = list.size();
+        Registries.ENTITY_TYPE.forEach(entityType -> list.add(entityType.toString()));
+        list.remove(EntityType.PLAYER.toString()); // Don't add players to the blacklist
+        SolidMobsMain.saveConfig();
+        SolidMobsMain.resetExemptions();
+        context.getSource().getServer().getPlayerManager().getPlayerList().forEach(SolidMobsCrossPlatformHelper::sendConfigToClient);
+        sendCommandFeedback(context, ((list.size() - size) + " new entities have been added to the " + name));
+        return 1;
+    }
+
+    private static @NotNull Command<ServerCommandSource> blacklistList() {
+        return c -> {
+            sendCommandFeedback(c, ("The HARDCODED collision blacklist (set by the settings or the dev): "
+                    + SolidMobsMain.EXEMPT_ENTITIES.toString()
+                    + "\n§7_____________________________§r\nThe Custom collision blacklist (set by you): "
+                    + SolidMobsMain.solidMobsConfigData.entityCollisionBlacklist));
+            return 1;
+        };
+    }
+
+    private static @NotNull Command<ServerCommandSource> bounceListList() {
+        return c -> {
+            sendCommandFeedback(c, ("The Custom entity bounce list: "
+                    + SolidMobsMain.solidMobsConfigData.entityBounceList));
+            return 1;
+        };
     }
 }